TFP-4570 rydding og forberede eget inngangsvilkaar-regel-repository
