+++ domenetjenester/inngangsvilkar/regelmodell/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/regelmodell/RegelUtfall.java
@@ -1,0 +1,6 @@
+package no.nav.foreldrepenger.inngangsvilkaar.regelmodell;
+
+public enum RegelUtfall {
+    OPPFYLT, IKKE_OPPFYLT, IKKE_VURDERT;
+}
+
