+++ domenetjenester/inngangsvilkar/regelmodell/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/regelmodell/RegelYtelse.java
@@ -1,0 +1,6 @@
+package no.nav.foreldrepenger.inngangsvilkaar.regelmodell;
+
+public enum RegelYtelse {
+    ENGANGSTØNAD, FORELDREPENGER, SVANGERSKAPSPENGER
+}
+
