+++ domenetjenester/inngangsvilkar/regelmodell/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/regelmodell/medlemskap/SjekkBrukerErAvklartSomIkkeMedlem.java
@@ -1,10 +1,9 @@
 package no.nav.foreldrepenger.inngangsvilkaar.regelmodell.medlemskap;
 
-import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.RegelUtfallMerknad;
 import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.MerknadRuleReasonRef;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.RegelUtfallMerknad;
 import no.nav.fpsak.nare.doc.RuleDocumentation;
 import no.nav.fpsak.nare.evaluation.Evaluation;
-import no.nav.fpsak.nare.evaluation.RuleReasonRef;
 import no.nav.fpsak.nare.specification.LeafSpecification;
 
 @RuleDocumentation(SjekkBrukerErAvklartSomIkkeMedlem.ID)
@@ -22,7 +21,7 @@ public class SjekkBrukerErAvklartSomIkkeMedlem extends LeafSpecification<Medlems
     @Override
     public Evaluation evaluate(MedlemskapsvilkårGrunnlag grunnlag) {
         if (!grunnlag.brukerErMedlem()) {
-            return ja(IKKE_OPPFYLT_BRUKER_ER_OPPFØRT_SOM_IKKE_MEDLEM);
+            return ja();
         }
         return nei();
     }
