+++ domenetjenester/inngangsvilkar/tjeneste/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/adopsjon/InngangsvilkårEngangsstønadAdopsjon.java
@@ -6,32 +6,33 @@ import javax.inject.Inject;
 import no.nav.foreldrepenger.behandling.BehandlingReferanse;
 import no.nav.foreldrepenger.behandlingslager.behandling.vilkår.VilkårType;
 import no.nav.foreldrepenger.inngangsvilkaar.Inngangsvilkår;
+import no.nav.foreldrepenger.inngangsvilkaar.RegelResultatOversetter;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårData;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårTypeRef;
-import no.nav.foreldrepenger.inngangsvilkaar.impl.InngangsvilkårOversetter;
-import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.adopsjon.AdopsjonsvilkårEngangsstønad;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.InngangsvilkårRegler;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.RegelYtelse;
 
 @ApplicationScoped
 @VilkårTypeRef(VilkårType.ADOPSJONSVILKÅRET_ENGANGSSTØNAD)
 public class InngangsvilkårEngangsstønadAdopsjon implements Inngangsvilkår {
 
-    private InngangsvilkårOversetter inngangsvilkårOversetter;
+    private AdopsjonsvilkårOversetter adopsjonsvilkårOversetter;
 
     InngangsvilkårEngangsstønadAdopsjon() {
         // for CDI proxy
     }
 
     @Inject
-    public InngangsvilkårEngangsstønadAdopsjon(InngangsvilkårOversetter inngangsvilkårOversetter) {
-        this.inngangsvilkårOversetter = inngangsvilkårOversetter;
+    public InngangsvilkårEngangsstønadAdopsjon(AdopsjonsvilkårOversetter adopsjonsvilkårOversetter) {
+        this.adopsjonsvilkårOversetter = adopsjonsvilkårOversetter;
     }
 
     @Override
     public VilkårData vurderVilkår(BehandlingReferanse ref) {
-        var grunnlag = inngangsvilkårOversetter.oversettTilRegelModellAdopsjon(ref);
+        var grunnlag = adopsjonsvilkårOversetter.oversettTilRegelModellAdopsjon(ref);
 
-        var evaluation = new AdopsjonsvilkårEngangsstønad().evaluer(grunnlag);
+        var resultat = InngangsvilkårRegler.adopsjon(RegelYtelse.ENGANGSTØNAD, grunnlag);
 
-        return InngangsvilkårOversetter.tilVilkårData(VilkårType.ADOPSJONSVILKÅRET_ENGANGSSTØNAD, evaluation, grunnlag);
+        return RegelResultatOversetter.oversett(VilkårType.ADOPSJONSVILKÅRET_ENGANGSSTØNAD, resultat);
     }
 }
