+++ domenetjenester/inngangsvilkar/tjeneste/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/adopsjon/InngangsvilkårForeldrepengerAdopsjon.java
@@ -6,31 +6,32 @@ import javax.inject.Inject;
 import no.nav.foreldrepenger.behandling.BehandlingReferanse;
 import no.nav.foreldrepenger.behandlingslager.behandling.vilkår.VilkårType;
 import no.nav.foreldrepenger.inngangsvilkaar.Inngangsvilkår;
+import no.nav.foreldrepenger.inngangsvilkaar.RegelResultatOversetter;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårData;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårTypeRef;
-import no.nav.foreldrepenger.inngangsvilkaar.impl.InngangsvilkårOversetter;
-import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.adopsjon.AdopsjonsvilkårForeldrepenger;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.InngangsvilkårRegler;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.RegelYtelse;
 
 @ApplicationScoped
 @VilkårTypeRef(VilkårType.ADOPSJONSVILKARET_FORELDREPENGER)
 public class InngangsvilkårForeldrepengerAdopsjon implements Inngangsvilkår {
 
-    private InngangsvilkårOversetter inngangsvilkårOversetter;
+    private AdopsjonsvilkårOversetter adopsjonsvilkårOversetter;
 
     public InngangsvilkårForeldrepengerAdopsjon() {
     }
 
     @Inject
-    public InngangsvilkårForeldrepengerAdopsjon(InngangsvilkårOversetter inngangsvilkårOversetter) {
-        this.inngangsvilkårOversetter = inngangsvilkårOversetter;
+    public InngangsvilkårForeldrepengerAdopsjon(AdopsjonsvilkårOversetter adopsjonsvilkårOversetter) {
+        this.adopsjonsvilkårOversetter = adopsjonsvilkårOversetter;
     }
 
     @Override
     public VilkårData vurderVilkår(BehandlingReferanse ref) {
-        var grunnlag = inngangsvilkårOversetter.oversettTilRegelModellAdopsjon(ref);
+        var grunnlag = adopsjonsvilkårOversetter.oversettTilRegelModellAdopsjon(ref);
 
-        var evaluation = new AdopsjonsvilkårForeldrepenger().evaluer(grunnlag);
+        var resultat = InngangsvilkårRegler.adopsjon(RegelYtelse.FORELDREPENGER, grunnlag);
 
-        return InngangsvilkårOversetter.tilVilkårData(VilkårType.ADOPSJONSVILKARET_FORELDREPENGER, evaluation, grunnlag);
+        return RegelResultatOversetter.oversett(VilkårType.ADOPSJONSVILKARET_FORELDREPENGER, resultat);
     }
 }
