+++ domenetjenester/inngangsvilkar/tjeneste/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/medlemskap/FinnOmSøkerHarArbeidsforholdOgInntekt.java
@@ -1,4 +1,4 @@
-package no.nav.foreldrepenger.inngangsvilkaar.impl;
+package no.nav.foreldrepenger.inngangsvilkaar.medlemskap;
 
 import static java.util.stream.Collectors.toList;
