+++ domenetjenester/inngangsvilkar/tjeneste/src/main/java/no/nav/foreldrepenger/inngangsvilkaar/medlemskap/InngangsvilkårMedlemskap.java
@@ -6,32 +6,33 @@ import javax.inject.Inject;
 import no.nav.foreldrepenger.behandling.BehandlingReferanse;
 import no.nav.foreldrepenger.behandlingslager.behandling.vilkår.VilkårType;
 import no.nav.foreldrepenger.inngangsvilkaar.Inngangsvilkår;
+import no.nav.foreldrepenger.inngangsvilkaar.RegelResultatOversetter;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårData;
 import no.nav.foreldrepenger.inngangsvilkaar.VilkårTypeRef;
-import no.nav.foreldrepenger.inngangsvilkaar.impl.InngangsvilkårOversetter;
-import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.medlemskap.Medlemskapsvilkår;
+import no.nav.foreldrepenger.inngangsvilkaar.regelmodell.InngangsvilkårRegler;
 
 @ApplicationScoped
 @VilkårTypeRef(VilkårType.MEDLEMSKAPSVILKÅRET)
 public class InngangsvilkårMedlemskap implements Inngangsvilkår {
 
-    private InngangsvilkårOversetter inngangsvilkårOversetter;
+    private MedlemsvilkårOversetter medlemsvilkårOversetter;
 
     InngangsvilkårMedlemskap() {
         // for CDI proxy
     }
 
     @Inject
-    public InngangsvilkårMedlemskap(InngangsvilkårOversetter inngangsvilkårOversetter) {
-        this.inngangsvilkårOversetter = inngangsvilkårOversetter;
+    public InngangsvilkårMedlemskap(MedlemsvilkårOversetter medlemsvilkårOversetter) {
+        this.medlemsvilkårOversetter = medlemsvilkårOversetter;
     }
 
     @Override
     public VilkårData vurderVilkår(BehandlingReferanse ref) {
-        var grunnlag = inngangsvilkårOversetter.oversettTilRegelModellMedlemskap(ref);
+        var grunnlag = medlemsvilkårOversetter.oversettTilRegelModellMedlemskap(ref);
 
-        var evaluation = new Medlemskapsvilkår().evaluer(grunnlag);
+        var resultat = InngangsvilkårRegler.medlemskap(grunnlag);
 
-        return InngangsvilkårOversetter.tilVilkårData(VilkårType.MEDLEMSKAPSVILKÅRET, evaluation, grunnlag);
+        return RegelResultatOversetter.oversett(VilkårType.MEDLEMSKAPSVILKÅRET, resultat);
+
     }
 }
